move state manipulation to individual file
