+++ catalog/app/components/FileEditor/QuiltConfigEditor/QuiltSummarize/QuiltSummarize.tsx
@@ -1,12 +1,9 @@
 import cx from 'classnames'
 import invariant from 'invariant'
-import { nanoid } from 'nanoid'
 import * as React from 'react'
 import * as RRDom from 'react-router-dom'
 import { useDebounce } from 'use-debounce'
 import * as M from '@material-ui/core'
-
-import quiltSummarizeSchema from 'schemas/quilt_summarize.json'
 
 import type * as Summarize from 'components/Preview/loaders/summarize'
 import Skeleton from 'components/Skeleton'
@@ -15,181 +12,24 @@ import * as requests from 'containers/Bucket/requests'
 import * as Listing from 'containers/Bucket/Listing'
 import { useData } from 'utils/Data'
 import * as Dialogs from 'utils/GlobalDialogs'
-import { makeSchemaValidator } from 'utils/JSONSchema'
 import Log from 'utils/Logging'
 import StyledLink from 'utils/StyledLink'
 
 import type { QuiltConfigEditorProps } from '../QuiltConfigEditor'
 
-interface FileExtended extends Omit<Summarize.FileExtended, 'types'> {
-  isExtended: boolean
-  type?: Summarize.TypeExtended
-}
-
-interface Column {
-  id: string
-  file: FileExtended
-}
-
-interface Row {
-  id: string
-  columns: Column[]
-}
-
-interface Layout {
-  rows: Row[]
-}
-
-const pathToFile = (path: string): FileExtended => ({ path, isExtended: false })
-
-const emptyFile: FileExtended = pathToFile('')
-
-const createColumn = (file: FileExtended): Column => ({
-  id: nanoid(3),
-  file,
-})
-
-const createRow = (file: FileExtended): Row => ({
-  id: nanoid(3),
-  columns: [createColumn(file)],
-})
-
-const init = (payload?: Layout) => (): Layout =>
-  payload || {
-    rows: [createRow(emptyFile)],
-  }
-
-function insert<T>(array: T[], index: number, item: T): T[] {
-  return array.toSpliced(index, 0, item)
-}
-
-function insertAfter<T extends { id: string }>(array: T[], id: string, item: T): T[] {
-  const index = array.findIndex((r) => r.id === id)
-  return insert(array, index + 1, item)
-}
-
-type Callback<T> = (item: T) => T
-function replace<T extends { id: string }>(array: T[], id: string, cb: Callback<T>): T[] {
-  const index = array.findIndex((r) => r.id === id)
-  return array.toSpliced(index, 1, cb(array[index]))
-}
-
-const addRowAfter =
-  (rowId: string) =>
-  (layout: Layout): Layout => ({
-    rows: insertAfter(layout.rows, rowId, createRow(emptyFile)),
-  })
-
-const addColumn =
-  (rowId: string, columnId: string) =>
-  (file: FileExtended) =>
-  (layout: Layout): Layout => ({
-    rows: replace(layout.rows, rowId, (row) => ({
-      ...row,
-      columns: insertAfter(row.columns, columnId, createColumn(file)),
-    })),
-  })
-
-const changeValue =
-  (rowId: string, columnId: string) =>
-  (file: Partial<FileExtended>) =>
-  (layout: Layout): Layout => ({
-    rows: replace(layout.rows, rowId, (row) => ({
-      ...row,
-      columns: replace(row.columns, columnId, (column) => ({
-        ...column,
-        file: {
-          ...column.file,
-          ...file,
-        },
-      })),
-    })),
-  })
-
-const removeColumn =
-  (rowId: string, columnId: string) =>
-  (layout: Layout): Layout => {
-    const rowIndex = layout.rows.findIndex((r) => r.id === rowId)
-    if (layout.rows[rowIndex].columns.length === 1) {
-      return {
-        rows: layout.rows.toSpliced(rowIndex, 1),
-      }
-    }
-    return {
-      rows: replace(layout.rows, rowId, (row) => ({
-        ...row,
-        columns: row.columns.filter((c) => c.id !== columnId),
-      })),
-    }
-  }
-
-function parseColumn(fileOrPath: Summarize.File): Column {
-  if (typeof fileOrPath === 'string') {
-    return createColumn(pathToFile(fileOrPath))
-  }
-  const { types, ...file } = fileOrPath
-  if (!types || !types.length) return createColumn({ ...fileOrPath, isExtended: true })
-  return createColumn({
-    ...file,
-    isExtended: true,
-    type: typeof types[0] === 'string' ? { name: types[0] } : types[0],
-  })
-}
-
-function preStringifyColumn(column: Column): Summarize.File {
-  const {
-    file: { isExtended, type, path, ...file },
-  } = column
-  if (!type) {
-    if (!Object.keys(file).length) return path
-    return {
-      path,
-      ...file,
-    }
-  }
-  return {
-    types: [type],
-    path,
-    ...file,
-  }
-}
-
-function validate(config: any) {
-  const errors = makeSchemaValidator(quiltSummarizeSchema)(config)
-  if (errors.length) {
-    throw new Error(errors.map((e) => e.message).join('\n'))
-  }
-  return undefined
-}
-
-function parse(str: string): Layout {
-  const config = JSON.parse(str)
-
-  if (!config) return { rows: [] }
-  if (!Array.isArray(config)) throw new Error('Expected array')
-
-  validate(config)
-
-  return {
-    rows: config.map((row) => ({
-      id: nanoid(2),
-      columns: Array.isArray(row) ? row.map(parseColumn) : [parseColumn(row)],
-    })),
-  }
-}
-
-function stringify(layout: Layout) {
-  const converted = layout.rows
-    .map((row) => {
-      const columns = row.columns.filter(({ file }) => file.path).map(preStringifyColumn)
-      return columns.length > 1 ? columns : columns[0]
-    })
-    .filter(Boolean)
-
-  validate(converted)
-
-  return JSON.stringify(converted, null, 2)
-}
+import {
+  addColumn,
+  addRowAfter,
+  changeValue,
+  emptyFile,
+  init,
+  parse,
+  removeColumn,
+  schema,
+  stringify,
+  useState,
+} from './State'
+import type { Column, FileExtended, Row, Layout } from './State'
 
 function useFormattedListing(r: requests.BucketListingResult): Listing.Item[] {
   return React.useMemo(() => {
@@ -565,7 +405,7 @@ function AddColumn({ className, column, disabled, last, onChange, row }: AddColu
                     <M.MenuItem value="">
                       <i>Default</i>
                     </M.MenuItem>
-                    {quiltSummarizeSchema.definitions.typeShorthand.enum.map((type) => (
+                    {schema.definitions.typeShorthand.enum.map((type) => (
                       <M.MenuItem key={type} value={type}>
                         {type}
                       </M.MenuItem>
@@ -822,7 +662,7 @@ export default function QuiltSummarize({
   onChange,
 }: QuiltConfigEditorProps) {
   const classes = useStyles()
-  const [layout, setLayout] = React.useState<Layout>(init())
+  const { layout, setLayout } = useState()
   const [state, setState] = React.useState<Error | null>(error)
 
   React.useEffect(() => {
@@ -832,7 +672,7 @@ export default function QuiltSummarize({
     } catch (e) {
       setState(e instanceof Error ? e : new Error(`${e}`))
     }
-  }, [initialValue])
+  }, [initialValue, setLayout])
 
   const [value] = useDebounce(layout, 300)
   React.useEffect(() => {
